Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/browsers/history/joint-session-history/joint-session-history-remove-iframe-expected.txt

CONSOLE MESSAGE: line 2659: Error: assert_equals: expected 2 but got 3


Harness Error (FAIL), message = Error: assert_equals: expected 2 but got 3

TIMEOUT Joint session history length does not include entries from a removed iframe. Test timed out
